rollback
I going to have to reimplement this a lot differently. This is due to a bug that if there are multiple overlapping items, the item after that may flag an address as a gap because it checks the ending address of the item before it instead of the one with the largest number of bytes that overlaps.

Here is a test:
7E0000	1	test1
7E0060	-7E0063	test2
7E0062	1	test3
7E0064	9	test4

results in 5th item on the output table, when sorted ascending, saying there is a 1-byte gap at $7E0063, despite the first overlap saying that $60-$63 is occupied.