update
Fixed the gap indicator, you actually need to have a variable that stores the highest address (so far) that is before the current [Index_WhichEntry]. Everytime you have an end address that is bigger, then replace whatever value in this variable to find the TRUE ending address.

The same should also apply to overlapped address, causing the overlap indicator to say that it isn't overlapping if the index before it have a lower ending address than the item before that.